requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cors==1.0.1
Django==4.1.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
filelock==3.12.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
inflection==0.5.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
psycopg2==2.9.6
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.28.2
requests-file==1.5.1
rpds-py==0.8.10
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
zope.event==4.6
zope.interface==6.0
django-environ==0.11.2